+++ changelog.md
@@ -1,8 +1,13 @@
-Version: ? -------------------------
+# Version: 1.2.1
+Changed from downloading segment files, then process to download and process on the go, only save the dataframe
+Changed naming convention of changelog.md
+
+# Version: 1.1.2 
 Update logging to be more consistent
 
-Version: ? -------------------------
+# Version: 1.1.1
 Changed it so directory paths are stored in the class instantiation
 
-Version: Prehistory ----------------
+# Version: 1.0 - Prehistory
 Nathan has done good working merging in our earlier changes
+